requirements.txt
Django==2.2.10
django-nested-admin==3.3.3
djangorestframework==3.12.2
psycopg2==2.8.6
python-monkey-business==1.0.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1

